test(worker): split runtime boundary cases
