+++ docs/REPO_MAP.md
@@ -256,7 +256,7 @@ does not establish same-user tamper resistance.
 
 | Path | Responsibility | Owning offline suite |
 |---|---|---|
-| `agy-worker.sh`, `skills/agy-worker/runtime/agy-worker.sh`, `skills/agy-worker/runtime/scripts/agy_dispatch.py`, `skills/agy-worker/runtime/scripts/agy_dispatch_worktree.py` | Root compatibility entry plus caller-owned selection, optional personas, workflow resolution (`explore`/`task`: 1..2, default 2; `project`: 1..5, default 5), private prompt/log staging, process-owning progress-aware dispatch, and a path-bound sibling engine for bounded no-follow Git/worktree observations. Current V9 candidate/lifecycle writes use explicit semantic-v1 candidate snapshots plus a separate stable root/Git boundary (canonical root dev/inode, marker, git/common dirs, format, and top-level), current-bound-only candidate-SHA Verification v2, and a driver-invoked verification-copy that rebinds result/schema/root/candidate before and after a no-follow isolated copy. It preserves regular bytes/executable bits, rebases every contained symlink inside the copy, and rejects broken/outward/Git-admin links (no `.git`) so writable driver checks do not reconcile ignored drift into the candidate. Wrapper parse errors remain `64`; post-parse copy runtime/binding/destination failures map to `20`. The bounded owner-controlled copy is not same-UID tamper resistance. V1 remains read-only; V3/V4 require first-transition SHA+rebound-migration approval and do not advertise the non-migrating copy helper. Finalization remains exact Codex-declared assurance after current artifact/schema/root/worktree rebinding; preserved provider-error/cancelled candidates, queued SHA+entry rebinding, and exact quota-terminal mapping remain unchanged. | `tests/test-agy-worker.sh` (331 cases); `tests/test-agy-worker-remediation.py` (87 focused cases); its non-discoverable case module is loaded by that canonical suite |
+| `agy-worker.sh`, `skills/agy-worker/runtime/agy-worker.sh`, `skills/agy-worker/runtime/scripts/agy_dispatch.py`, `skills/agy-worker/runtime/scripts/agy_dispatch_worktree.py` | Root compatibility entry plus caller-owned selection, optional personas, workflow resolution (`explore`/`task`: 1..2, default 2; `project`: 1..5, default 5), private prompt/log staging, process-owning progress-aware dispatch, and a path-bound sibling engine for bounded no-follow Git/worktree observations. Current V9 candidate/lifecycle writes use explicit semantic-v1 candidate snapshots plus a separate stable root/Git boundary (canonical root dev/inode, marker, git/common dirs, format, and top-level), current-bound-only candidate-SHA Verification v2, and a driver-invoked verification-copy that rebinds result/schema/root/candidate before and after a no-follow isolated copy. It preserves regular bytes/executable bits, rebases every contained symlink inside the copy, and rejects broken/outward/Git-admin links (no `.git`) so writable driver checks do not reconcile ignored drift into the candidate. Wrapper parse errors remain `64`; post-parse copy runtime/binding/destination failures map to `20`. The bounded owner-controlled copy is not same-UID tamper resistance. V1 remains read-only; V3/V4 require first-transition SHA+rebound-migration approval and do not advertise the non-migrating copy helper. Finalization remains exact Codex-declared assurance after current artifact/schema/root/worktree rebinding; preserved provider-error/cancelled candidates, queued SHA+entry rebinding, and exact quota-terminal mapping remain unchanged. | `tests/test-agy-worker.sh` (331 cases); `tests/test-agy-worker-remediation.py` (87 focused cases); its non-discoverable case modules are loaded by that canonical suite |
 | `model-selection.sh`, `skills/agy-worker/runtime/model-selection.sh`, `skills/agy-worker/runtime/scripts/model_selection.py`, portable matrix/schema/SHA | Root compatibility entry plus exact matrix-bound model/effort resolution, CLI-only version-independent literal records, bounded safe-target semantic version plus structural critical-help preflight, automatic V2 mechanical launch authority for an exact version match, and explicit Codex disposition plus exact raw-help SHA for drift. V3 records bind that drift decision, capabilities/matrix/selection facts, and a bounded no-follow descriptor digest plus complete safe executable path authority (only the macOS `/var` alias is normalized); controller help prose is data, never availability inference. Codex inspects current bounded raw help before every reviewed direct dispatch and owns the semantic stop decision, including for an exact-version match. Literal version observation remains non-gating and selection provenance is driver-owned. | dispatcher, doctor, and packaging suites |
 | `model-recommendation.sh`, `skills/agy-worker/runtime/model-recommendation.sh`, `skills/agy-worker/runtime/scripts/model-recommendation.py` | Root compatibility entry plus side-effect-free pre/post recommendations; direct selections are labelled but unranked and never applied | `tests/test-agy-worker.sh` (331 cases) |
 | `doctor.sh`, `skills/agy-worker/runtime/doctor.sh`, `skills/agy-worker/runtime/scripts/doctor-metadata.py`, `skills/agy-worker/runtime/compat/` | Root compatibility entry plus deterministic offline prerequisite checks and byte-synchronized portable agy metadata | `tests/test-doctor.sh` (257 cases) plus packaging synchronization checks |
